solve linter errors and work on reformating
code
