+++ src/Components/QuestionsAndAnswers/QuestionsAndAnswers.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
+// import helpers from '../../Shared/api';
 // require helpers from '../../shared/api.js'
 // not sure how to import info from api calls
 
 const QuestionsAndAnswers = () => (
-    <div className="questions-and-answers">QuestionsAndAnswers PLACEHOLDER</div>
-
+  <div className="questions-and-answers">
+    {/* {helpers.getQuestions(2)} */}
+    QuestionsAndAnswers PLACEHOLDER
+  </div>
 );
 
 export default QuestionsAndAnswers;
